[GSOC] Build a single 'main' test program. This main program will be linked with all libraries in similar fashion as xbmc.bin and libxbmc.so.
This required modifications to the build system so that the main() function
for xbmc.bin does not get linked with the main() function for the
testsuite program. This has the side effect of allowing xbmc.bin and
libxbmc.so to be built from a single build.
