+++ xbmc/main/Makefile
@@ -1,0 +1,7 @@
+SRCS= \
+  main.cpp
+
+LIB=main.a
+
+include ../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
