+++ xbmc/test/Makefile
@@ -1,0 +1,9 @@
+SRCS=	\
+	xbmc-test.cpp
+
+LIB=xbmc-test.a
+
+INCLUDES += -I../../lib/gtest/include
+
+include ../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
