--- xbmc/threads/test/TestMain.cpp
@@ -1,49 +1,0 @@
-/*
- *      Copyright (C) 2005-2011 Team XBMC
- *      http://www.xbmc.org
- *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, write to
- *  the Free Software Foundation, 675 Mass Ave, Cambridge, MA 02139, USA.
- *  http://www.gnu.org/copyleft/gpl.html
- *
- */
-
-#include "TestHelpers.h"
-
-#include "threads/Thread.h"
-#include "commons/ilog.h"
-
-using ::testing::InitGoogleTest;
-
-class NullLogger : public XbmcCommons::ILogger
-{
-public:
-  void log(int loglevel, const char* message) {}
-};
-
-int main(int argc, char **argv)
-{
-  InitGoogleTest(&argc, argv);
-
-  // we need to configure CThread to use a dummy logger
-  NullLogger* nullLogger = new NullLogger();
-  CThread::SetLogger(nullLogger);
-
-  int ret = RUN_ALL_TESTS();
-
-  delete nullLogger;
-
-  return ret;
-}
-
